% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{enrichment}
\alias{enrichment}
\title{Compute the overlap}
\usage{
enrichment(self, ...)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{...}{The rest of the parameters}
}
\value{
the enrichment object
}
\description{
See 'enrichment.rrho' for full documentation.
}
