% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{enrichment.rrho}
\alias{enrichment.rrho}
\title{Compute the overlap}
\usage{
\method{enrichment}{rrho}(self, i, j, directions = "downdown", ...)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{i}{is the y coordinates of the rectangle (in the 'b' vector)}

\item{j}{is the x coordinates of the rectangle (in the 'a' vector)}

\item{directions}{is the direction in both list (a, b): "upup", "downup", "updown" or "downdown"}

\item{...}{The rest of the parameters}
}
\value{
the enrichment
}
\description{
Compute the overlap
}
