% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{ggplot.rrho}
\alias{ggplot.rrho}
\title{Compatibility function with old version (do NOT use in new software)}
\usage{
\method{ggplot}{rrho}(
  data = NULL,
  mapping = ggplot2::aes(),
  ...,
  environment = parent.frame()
)
}
\arguments{
\item{data}{the data}

\item{mapping}{the mapping}

\item{...}{all the parameters}

\item{environment}{the environment}
}
\value{
A ggplot overlap map
}
\description{
Compatibility function with old version (do NOT use in new software)
}
