% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{ggRedRibbon}
\alias{ggRedRibbon}
\title{Plot the overlap}
\usage{
ggRedRibbon(self, ...)
}
\arguments{
\item{self}{the RedRibbon object}

\item{...}{The rest of the parameters}
}
\value{
a ggplot object
}
\description{
see 'ggRedRibbon.rrho' for full documentation
}
