% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{newFC}
\alias{newFC}
\title{Create a fold change prediction object}
\usage{
newFC(deps, beta)
}
\arguments{
\item{deps}{is a vector of integer referring to the feature used for the predicton.
The values should be either -1 for an unpredicted features or in 1:length(deps).}

\item{beta}{is a vector of double. The fold change will be predicted as FCy = beta * FCx.}
}
\value{
The fold change prediction object
}
\description{
Create a fold change prediction object
}
