% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{newLDFIT}
\alias{newLDFIT}
\title{Create a correlation S3 object}
\usage{
newLDFIT(position, deps, half = 6480.306)
}
\arguments{
\item{position}{is the position vector}

\item{deps}{is the dependency vector}

\item{half}{is the half distance for the fitted function}
}
\value{
a correlation object
}
\description{
Create a correlation S3 object
}
