% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{newP}
\alias{newP}
\title{Create a correlation S3 object}
\usage{
newP(deps, r)
}
\arguments{
\item{deps}{is the dependency vector}

\item{r}{is the correlation coefficient}
}
\value{
a correlation object for P-Values
}
\description{
Create a correlation S3 object
}
