% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{permutation}
\alias{permutation}
\title{Compute the adjusted P-value in a rectangle.}
\usage{
permutation(self, ...)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{...}{The rest of the parameters}
}
\value{
the adjusted p-value
}
\description{
See 'permutation.rrho' for full documentation.
}
