% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{permutation.rrho}
\alias{permutation.rrho}
\title{Compute the adjusted P-value in a rectangle.}
\usage{
\method{permutation}{rrho}(
  self,
  i,
  j,
  i.len,
  j.len,
  a,
  b,
  algo_params = NULL,
  direction = "enrichment",
  algorithm = "classic",
  correlation = NULL,
  niter = 96,
  pvalue_i,
  pvalue_j,
  ...
)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{i}{is the y coordinates of the rectangle (in the 'b' vector)}

\item{j}{is the x coordinates of the rectangle (in the 'a' vector)}

\item{i.len}{is the vertical length of the rectangle}

\item{j.len}{is the horizontal length of the rectangle}

\item{a}{is the first list}

\item{b}{is the second list}

\item{algo_params}{is a list if 2 elements named m and n representing the number of P-value to compute between in the rectangle (only used for 'algorithm="classic"')}

\item{direction}{is the enrichment mode if set to 'enrichment' search for enrichment, otherwise search for underenrichment}

\item{algorithm}{is the algorithm used either 'classic' grid method or 'ea' evolutionary algorithm}

\item{correlation}{is the correlation between the genes}

\item{niter}{is the number of permutation}

\item{pvalue_i}{is the y coordinate of the best P-value}

\item{pvalue_j}{is the x coordinate of the best P-value}

\item{...}{The rest of the parameters}
}
\value{
a permutation
}
\description{
Compute the adjusted P-value in a rectangle.
}
