% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{quadrants}
\alias{quadrants}
\title{Compute the best coordinates of a RedRibbon S3 object.}
\usage{
quadrants(self, ...)
}
\arguments{
\item{self}{is a RedRibbon object created by 'RedRibbon' constructor}

\item{...}{The rest of the parameters}
}
\value{
the overlap for each quadrant
}
\description{
See 'quadrants.rrho' for full documentation.
}
