% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{quadrants.rrho}
\alias{quadrants.rrho}
\title{Compute the best coordinates of a RedRibbon S3 object.}
\usage{
\method{quadrants}{rrho}(
  self,
  m = NULL,
  n = NULL,
  whole = TRUE,
  whole.fraction = 1,
  algorithm = "ea",
  permutation = FALSE,
  niter = 96,
  ...
)
}
\arguments{
\item{self}{is a RedRibbon object created by 'RedRibbon' constructor}

\item{m}{is the number of coordinates to compute on the y axis (b)}

\item{n}{is the number of coordinates to compute on the x axis (a)}

\item{whole}{if TRUE run the whole list otherwise run by quadrants.}

\item{whole.fraction}{is the fraction of the quadrant analysed in 'whole' mode.}

\item{algorithm}{is the algorithm used to find the minimal p-value: 
\itemize{

\item{"classic"} {Grid algorithm.} 
\item{"ea"} {Evolutionary algorithm for the greatest accuracy (Default)}.
}}

\item{permutation}{is TRUE if the permutation mode is run.}

\item{niter}{is the number of iteration for the permutation mode.}

\item{...}{The rest of the parameters}
}
\value{
A list of coordinates, one by quadrant.
}
\description{
Compute the best coordinates of a RedRibbon S3 object.
}
\examples{
library(RedRibbon)

a <- as.double(1:1000) - 450
b <- as.double(1:1000) - 460
rr <- RedRibbon(a, b)

quad <- quadrants(rr, m=1000, n=1000)

}
