% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{rectangle_min}
\alias{rectangle_min}
\title{Search for the minimal P-value enrichment in a rectangle}
\usage{
rectangle_min(self, ...)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{...}{The rest of the parameters}
}
\value{
the minimum in a rectangle area
}
\description{
See 'rectangle_min.rrho' for full documentation.
}
