% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{rectangle_min.rrho}
\alias{rectangle_min.rrho}
\title{Search for the minimal P-value enrichment in a rectangle}
\usage{
\method{rectangle_min}{rrho}(
  self,
  i,
  j,
  i.len,
  j.len,
  m = NULL,
  n = NULL,
  direction = "enrichment",
  algorithm = "classic",
  ...
)
}
\arguments{
\item{self}{is a RedRibbon S3 object}

\item{i}{is the y coordinates of the rectangle (in the 'b' vector)}

\item{j}{is the x coordinates of the rectangle (in the 'a' vector)}

\item{i.len}{is the vertical length of the rectangle}

\item{j.len}{is the horizontal length of the rectangle}

\item{m}{is the number of P-values to compute between (i, .) and (i + i.len, .) (only used for 'algorithm="classic"')}

\item{n}{is the number of P-values to compute between (., j) and (., j + j.lenn)}

\item{direction}{is the enrichment mode if set to 'enrichment' search for enrichment, otherwise search for underenrichment}

\item{algorithm}{is the algorithm used either 'classic' grid method or 'ea' evolutionary algorithm}

\item{...}{The rest of the parameters}
}
\value{
the minimum in a rectangle
}
\description{
Search for the minimal P-value enrichment in a rectangle
}
