% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{RedRibbon.data.frame}
\alias{RedRibbon.data.frame}
\title{Creates a RedRibbon object from 2 numeric vectors}
\usage{
\method{RedRibbon}{data.frame}(
  self,
  enrichment_mode = c("hyper", "hyper-two-tailed", "hyper-two-tailed-old"),
  correlation = NULL,
  ...
)
}
\arguments{
\item{self}{a data.frame. The real number columns a and b are mandatory.}

\item{enrichment_mode}{\itemize{
   \item{"hyper"} {for one tailed hypergeometric test}
   \item{"hyper-two-tailed"} {for one tailed hypergeometric test}
   \item{"hyper-two-tailed-old"} {for the original R package two tailed test.}
}}

\item{correlation}{the correlation vector (see newFC function).}

\item{...}{The rest of the parameters}
}
\value{
A RedRibbon S3 object.
}
\description{
Creates a RedRibbon object from 2 numeric vectors
}
\examples{

library(RedRibbon)
rr <- RedRibbon(data.frame(id=c("gene1", "gene2", "gene3", "gene4"),
                           a=c(0.5, 0.7,0.3, 0.8),
                           b=c(0.6,0.6,0.4,0.7)))

}
