% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{RedRibbon.numeric}
\alias{RedRibbon.numeric}
\title{Creates a RedRibbon objects from 2 numeric vectors}
\usage{
\method{RedRibbon}{numeric}(self, b, ...)
}
\arguments{
\item{self}{is a vector of double.}

\item{b}{is a vector of double.}

\item{...}{see documentation of 'RedRibbon.data.frame'}
}
\value{
A rrho S3 object.
}
\description{
A helper function to check that \code{a} and \code{b} input parameters are of the same length.
The function is equivalent to RedRibbon(data.frame(a=a, b=b), ...) (see ).
}
\examples{

library(RedRibbon)
rr <- RedRibbon(c(0.5, 0.7,0.3, 0.8), c(0.6,0.6,0.4,0.7))

}
