% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{RedRibbon}
\alias{RedRibbon}
\title{RedRibbon S3 object constructor}
\usage{
RedRibbon(self, ...)
}
\arguments{
\item{self}{a data.frame or vector}

\item{...}{The rest of the parameters}
}
\value{
a RedRibbon S3 object
}
\description{
Pass a 'data.frame or a two vectors of numerics to call either 'RedRibbon.data.frame' or 'RedRibbon.numeric'.
}
