% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{RRHO}
\alias{RRHO}
\title{Drop in replacement for original R function}
\usage{
RRHO(
  list1,
  list2,
  stepsize = NULL,
  labels = c("a", "b"),
  alternative = c("two.sided", "enrichment"),
  plots = FALSE,
  outputdir = NULL,
  BY = FALSE,
  log10.ind = FALSE
)
}
\arguments{
\item{list1}{is the first list}

\item{list2}{is the second list}

\item{stepsize}{is the step size}

\item{labels}{are the labels}

\item{alternative}{is 'two.sided' or  'enrichment'}

\item{plots}{is a flag to plot}

\item{outputdir}{is the output directory}

\item{BY}{is the correction}

\item{log10.ind}{is the logarithm of the hypermat}
}
\value{
the a list(rr = ..., quadrants = ...) with 'rr' the RedRibbon object and 'quadrants' the result of the overlap
}
\description{
'RRHO' function preserve the compatibility with the original package but
with the performance and accuracy of the new method. The stepsize parameter is
used only to return the hypermat and ignored to determine the minimal p-values
in the four quadrants as the evolutionary algorithm is used.
}
