% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_expression_prediction}
\alias{rrho_expression_prediction}
\title{Compute dependencies between genes}
\usage{
rrho_expression_prediction(mat, nbr.tested = -1)
}
\arguments{
\item{mat}{an expression matrix}

\item{nbr.tested}{limit the number of tested correlation}
}
\value{
list(index=..., beta=...)
}
\description{
See vignette for examples.
}
