% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_ij}
\alias{rrho_ij}
\title{Call to C function (internal use): RedRibbon rrho_ij}
\usage{
rrho_ij(i, j, a, b, mode = c("hyper"))
}
\arguments{
\item{i}{is the y coordinates of the rectangle (in the `b` vector)}

\item{j}{is the x coordinates of the rectangle (in the `a` vector)}

\item{a}{is the first list}

\item{b}{is the second list}

\item{mode}{is the enrichment mode: "hyper", "hyper-two-tailed" or "hyper-two-tailed-old"}
}
\value{
the hypergeometric p-value
}
\description{
Call to C function (internal use): RedRibbon rrho_ij
}
