% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_intersect}
\alias{rrho_intersect}
\title{Call to C function (internal use): Compute the overlap}
\usage{
rrho_intersect(i, j, a, b, directions = c("downdown"))
}
\arguments{
\item{i}{is the y coordinates of the rectangle (in the `b` vector)}

\item{j}{is the x coordinates of the rectangle (in the `a` vector)}

\item{a}{is the first list}

\item{b}{is the second list}

\item{directions}{is the direction in both list (a, b): "upup", "downup", "updown" or "downdown"}
}
\value{
the overlap
}
\description{
Call to C function (internal use): Compute the overlap
}
