% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_ldfit_prediction}
\alias{rrho_ldfit_prediction}
\title{RedRibbon rrho_ldfit_prediction}
\usage{
rrho_ldfit_prediction(half, pval, position)
}
\arguments{
\item{half}{is the half distance for the fitted function}

\item{pval}{are the pvalues}

\item{position}{is the position vector}
}
\value{
the approximate dependencies
}
\description{
RedRibbon rrho_ldfit_prediction
}
\examples{

library(RedRibbon)
deps <- rrho_ldfit_prediction(600, runif(100), 1:100)

}
