% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_normalize}
\alias{rrho_normalize}
\title{Normalize TPMs}
\usage{
rrho_normalize(
  mat,
  ref,
  mode = c("poisson", "ls_mean", "ls_variance", "geometric_mean")
)
}
\arguments{
\item{mat}{is a matrix of expression}

\item{ref}{are the index of the reference}

\item{mode}{"poisson", "ls_mean", "ls_variance", "geometric_mean"}
}
\value{
a normalized expression matrix
}
\description{
Normalize TPMs
}
