% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{rrho_rectangle}
\alias{rrho_rectangle}
\title{Call to C function (internal use): compute the full grid rectangle}
\usage{
rrho_rectangle(i, j, i.len, j.len, m, n, a, b, mode = c("hyper"), LOG = FALSE)
}
\arguments{
\item{i}{is the y coordinates of the rectangle (in the `b` vector)}

\item{j}{is the x coordinates of the rectangle (in the `a` vector)}

\item{i.len}{is the vertical length of the rectangle}

\item{j.len}{is the horizontal length of the rectangle}

\item{m}{is the number of P-values to compute between (i, .) and (i + i.len, .)}

\item{n}{is the number of P-values to compute between (., j) and (., j + j.len)}

\item{a}{is the first list}

\item{b}{is the second list}

\item{mode}{is the enrichment mode: "hyper", "hyper-two-tailed" or "hyper-two-tailed-old"}

\item{LOG}{is a flag if set return -log pval}
}
\value{
The hypergeometric p-values for a rectangle
}
\description{
Call to C function (internal use): compute the full grid rectangle
}
