% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{setoptions}
\alias{setoptions}
\title{Set RedRibbon options on a RedRibbon S3 object}
\usage{
setoptions(self, ...)
}
\arguments{
\item{self}{a RedRibbon S3 object}

\item{...}{The rest of the parameters}
}
\value{
a RedRibbon S3 object
}
\description{
See 'setoptions.rrho' for full documentation.
}
