% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho.R
\name{setoptions.rrho}
\alias{setoptions.rrho}
\title{Set RedRibbon options on a RedRibbon S3 object}
\usage{
\method{setoptions}{rrho}(self, enrichment_mode = NULL, ggplot_colours = NULL, ...)
}
\arguments{
\item{self}{A RedRibbon object.}

\item{enrichment_mode}{A string (default = NULL). 
   \itemize{
   \item{"enrichment"} {for one tailed hypergeometric test}
   \item{"hyper-two-tailed"} {for one tailed hypergeometric test, "hyper-two-tailed-old", for the original R package two tailed test.}
}}

\item{ggplot_colours}{is the color palette used for the plots. The default is
\code{
colfunc <- grDevices::colorRampPalette(c("#eb3434", "#eb9334", "#ebeb34", "#49eb34", "#34eba5", "#34b4eb", "#3446eb"))
colors  <-  colfunc(1000)
ggplot_colours = c(colors, rev(colors))
}}

\item{...}{The rest of the parameters}
}
\value{
A RedRibbon S3 object with updated parameters.
}
\description{
Helper function allowing to change the enrichment method that you wish to use as well as the color scheme
for the RedRibbon level map.
}
\examples{
library(magrittr)
library(RedRibbon)
RedRibbon(c(0.5, 0.7,0.3, 0.8), c(0.6,0.6,0.4,0.7)) \%>\% 
 setoptions(enrichment_mode="hyper-two-tailed")

}
